test(test_peaks_over_threshold): ensure that exceedance extraction provides the correct result
The function get_exceedance_peaks_over_threshold executes the function get_threshold_peaks_over_threshold to get pot thresholds before subtracting the values with the respective threshold. Now, the low anomaly_type is a bit tricky, but it is solved! The trick is to consider exceedances as always a positive value because it is the difference between data value and threshold value.