feat(pot): create functions that utilises Peaks Over Threshold method (#6)
* fix(pot_window): change name from pot_windows to pot_window

* feat(time_series): create time_series/ directory for uploading dataset

* build(toml): add opynpyxl when users install [all] features

* feat(time_series): create an upload feature for csv and excel

For lazy users that want to get directly the correct form of Series 😜

* feat(peaks_over_threshold): implement get_pot_threshold to calculate dynamic threshold for exceedance extraction in POT method

* feat(import-strategy): import all functions in peaks_over_threshold.py in anomalytics/stats/__init__.py

* test(get_pot_threshold): ensure that threshold is calculated dynamically for POT method

* feat(get_exceedance_peaks_over_threshold): create the function for exceedance extraction with POT method

* test(test_peaks_over_threshold): ensure that exceedance extraction provides the correct result

The function get_exceedance_peaks_over_threshold executes the function get_threshold_peaks_over_threshold to get pot thresholds before subtracting the values with the respective threshold. Now, the low anomaly_type is a bit tricky, but it is solved! The trick is to consider exceedances as always a positive value because it is the difference between data value and threshold value.

* feat(log): add logger for debugging

* refactor(test_peaks_over_threshold): restructure the import

* test(test_time_series): ensure the read_ts wrapper function initializes the correct pd.Series

read_ts is an overload and wrapper function for pd.read_csv and pd.read_excel.

* ci(ci): add "extra" when install edit the dependencies

---------

Co-authored-by: N. L <nino@pleno.earth>